Fix #12 Put telnet in separate File
